restructuring of frontend & a bit of backend
