+++ frontend/src/pages/HomePage.js
@@ -1,7 +1,7 @@
 import NoteCreateHomeForm from "../components/NoteCreateHomeForm";
 
 const HomePage = () => {
-    return <NoteCreateHomeForm />;
+    return <NoteCreateHomeForm/>;
 };
 
 export default HomePage;
